requirements.txt
streamlit
pandas
plotly
nfcpy
openpyxl
